Remove 443 has in doc
Depends if HTTPS or not enable but the detection not works with SSL, we have to create our registry directly and skip check (With FQDN:5000 the server is detected but it's on the host is self not a fully HTTPS check.). 

So go on the UI, and use SSL by checkbox and the name MUST BE THE SAME HAS THE HOSTNAME OF REGISTRY CONF so don't add any port, and skip change or change it after.

Look docker registry section.